app-streamlit/requirements.txt
streamlit
streamlit-aggrid
streamlit_option_menu
streamlit-feedback
streamlit-tags
st-pages
click
np_utils
matplotlib
seaborn
pandas
graphviz
xgboost
catboost
vecstack
numpy

plotly
scikit-learn
lightgbm
openpyxl
joblib

#“unstructured[md,pdf,pptx]”
unstructured-client
unstructured
unstructured-inference
unstructured.pytesseract
urllib3
python-dotenv
panel
ipython
python-pptx
pdf2image
pdfminer
opencv-python

pikepdf
pypdf
pyyaml
dash-ag-grid
pycountry
cartopy

# Deep Learning dependencies for image generation
torch
torchvision
torchaudio
Pillow
tqdm
diffusers
peft